refactor: (W-073) extract FormattedContent and HtmlFragment into shared module
Move FormattedContent, InlineFormat from Chat.tsx and HtmlFragment from
SeedChat.tsx into a shared FormattedContent.tsx module. Both consumers
now import from the shared module instead of defining locally.
