+++ web/src/components/Chat.tsx
@@ -1,6 +1,7 @@
 import { useState, type RefObject } from "react";
 import type { ChatMessage } from "../hooks/useChat";
 import { TypingIndicator } from "./ActivityIndicator";
+import { FormattedContent } from "./FormattedContent";
 import { api } from "../api/client";
 
 interface Props {
@@ -133,84 +134,6 @@ export default function Chat({ messages, onSend, onReset, bottomRef, connected, 
   );
 }
 
-// Box-drawing characters used in Claude Code TUI tables
-const BOX_CHARS = /[┌┐└┘├┤┬┴┼│─]/;
-// Markdown pipe tables: lines starting with | and containing at least one more |
-const MD_TABLE = /^\s*\|.*\|/;
-
-/**
- * Render message content with basic formatting:
- * - Box-drawing table blocks → monospace <pre>
- * - **bold** → <strong>
- * - `code` → <code>
- * - Newlines preserved
- */
-function FormattedContent({ text }: { text: string }) {
-  const lines = text.split("\n");
-  const blocks: { type: "text" | "table"; lines: string[] }[] = [];
-  let current: { type: "text" | "table"; lines: string[] } = { type: "text", lines: [] };
-
-  for (const line of lines) {
-    const isTable = BOX_CHARS.test(line) || MD_TABLE.test(line);
-    if (isTable && current.type !== "table") {
-      if (current.lines.length) blocks.push(current);
-      current = { type: "table", lines: [line] };
-    } else if (!isTable && current.type === "table") {
-      blocks.push(current);
-      current = { type: "text", lines: [line] };
-    } else {
-      current.lines.push(line);
-    }
-  }
-  if (current.lines.length) blocks.push(current);
-
-  return (
-    <>
-      {blocks.map((block, i) =>
-        block.type === "table" ? (
-          <pre
-            key={i}
-            className="my-1 text-[11px] leading-tight overflow-x-auto text-zinc-300 font-mono"
-          >
-            {block.lines.join("\n")}
-          </pre>
-        ) : (
-          <span key={i}>
-            {block.lines.map((line, j) => (
-              <span key={j}>
-                {j > 0 && <br />}
-                <InlineFormat text={line} />
-              </span>
-            ))}
-          </span>
-        )
-      )}
-    </>
-  );
-}
-
-/** Render inline formatting: **bold**, `code` */
-function InlineFormat({ text }: { text: string }) {
-  // Split on **bold** and `code` patterns
-  const parts = text.split(/(\*\*[^*]+\*\*|`[^`]+`)/g);
-  return (
-    <>
-      {parts.map((part, i) => {
-        if (part.startsWith("**") && part.endsWith("**")) {
-          return <strong key={i} className="font-semibold">{part.slice(2, -2)}</strong>;
-        }
-        if (part.startsWith("`") && part.endsWith("`")) {
-          return (
-            <code key={i} className="bg-zinc-800 px-1 py-0.5 rounded text-emerald-300 text-xs font-mono">
-              {part.slice(1, -1)}
-            </code>
-          );
-        }
-        return <span key={i}>{part}</span>;
-      })}
-    </>
-  );
-}
 
 function messageAlignment(source: string): string {
   if (source === "user") return "text-right";
